python/requirements.txt
pandas~=2.2.3
chess~=1.11.1
mysql-connector-python~=9.2.0
requests~=2.32.3